description_gen/requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
alabaster==0.7.12
altair==4.1.0
antlr4-python3-runtime==4.8
apyori==1.1.2
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.0
attrs==20.2.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
better-exceptions==0.3.3
bleach==4.1.0
blinker==1.4
blis==0.7.5
boto3==1.20.22
botocore==1.23.22
Brotli==1.0.9
cachetools==4.2.4
captum==0.4.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clean-text==0.5.0
click==7.1.2
colorama==0.4.4
coloredlogs==15.0.1
convokit==2.5.1
cryptography==37.0.2
cycler==0.11.0
cymem==2.0.6
datasets==1.16.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
detoxify==0.5.0
dill==0.3.4
distro==1.7.0
docformatter==1.4
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
emoji==1.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
extratools==0.8.2.1
fairscale==0.4.3
filelock==3.4.0
flake8==4.0.1
flake8-bugbear==21.11.29
Flask==2.0.2
Flask-Cors==3.0.10
flatbuffers==2.0
frozenlist==1.2.0
fsspec==2021.10.1
ftfy==6.0.3
future==0.18.2
gast==0.4.0
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.1.24
google-api-core==2.7.1
google-api-python-client==2.50.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.41.1
h5py==3.5.0
httplib2==0.20.4
huggingface-hub==0.2.1
humanfriendly==10.0
hydra-core==1.1.1
idna==3.3
imageio==2.19.3
imagesize==1.3.0
importlib-metadata==4.2.0
inflate64==0.1.3
iniconfig==1.1.1
iopath==0.1.9
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonlines==3.0.0
jsonschema==4.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
krippendorff==0.5.1
langcodes==3.3.0
libclang==12.0.0
loguru==0.5.3
lxml==4.9.0
Markdown==3.3.2
markdown-it-py==0.5.8
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
msgpack==1.0.3
msgpack-numpy==0.4.7.1
multidict==5.2.0
multiprocess==0.70.12.2
multivolumefile==0.2.3
murmurhash==1.0.6
myst-parser==0.12.10
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.3
nltk==3.6.5
notebook==6.4.5
numpy==1.21.0
oauthlib==3.1.1
omegaconf==2.1.1
openai==0.14.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.47
pandocfilters==1.5.0
parlai==1.6.0
parmap==1.5.3
parso==0.8.2
pathtools==0.1.2
pathy==0.6.1
patsy==0.5.2
pdfminer.six==20220524
pdftotree==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
pluggy==1.0.0
portalocker==2.3.2
prefixspan==0.5.2
preshed==3.0.6
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
py-gfm==1.0.2
py-rouge==1.1
py7zr==0.19.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==0.6.1
pycodestyle==2.8.0
pycparser==2.20
pycryptodomex==3.15.0
pydantic==1.8.2
pydeck==0.7.1
pyDeprecate==0.3.1
pyflakes==2.4.0
Pygments==2.10.0
PyMuPDF==1.19.6
pyparsing==2.4.7
PyPDF2==2.1.0
pyphen==0.12.0
pyppmd==0.18.3
pyrsistent==0.18.0
pytest==6.2.5
pytest-datadir==1.3.1
pytest-regressions==2.2.0
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-lightning==1.5.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.2
regex==2021.11.10
requests==2.26.0
requests-mock==1.9.3
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.46
scikit-image==0.19.2
scikit-learn==1.0.1

scipy==1.7.1
seaborn==0.11.2
selectivesearch==0.4
Send2Trash==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
sentry-sdk==1.5.5
sentsplit==1.0.7
sh==1.14.2
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.1
Sphinx==2.2.2
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.3
stanza==1.3.0
statsmodels==0.13.2
streamlit==1.1.0
subword-nmt==0.3.8
tabula-py==2.4.0
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
textstat==0.7.3
texttable==1.6.4
thinc==8.0.17
threadpoolctl==3.0.0
tifffile==2022.5.4
tokenizers==0.12.1
toml==0.10.2
tomli==1.2.3
toolz==0.11.1

torch-tb-profiler==0.3.1

torchmetrics==0.6.0
torchtext==0.11.0

tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.20
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.2
untokenize==0.1.1
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
Wand==0.6.7
wandb==0.12.10
wasabi==0.9.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websocket-server==0.6.2
Werkzeug==2.0.2
widgetsnbextension==3.5.2
word2number==1.1
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
